New das_query.py script
This script queries DAS to find the LFNs for the files of those samples
defined in samples_nosplit.ini.

The positional argument "sample_config" is the samples_nosplit.ini file
storing the query options and primary dataset names. The optional
argument "--outdir" is a name for the subdirectory created within the
configuration directory to hold the lists of LFNs. If not passed, the
LFN lists will be saved directly within the configuration directory.

Rather than importing the das_client as a module, the script calls the
das_client through a subprocess. This design choice was made because the
main() function in das_client.py does not take an argv argument, which
means taking advantage of its convenient features requires tampering
with sys.argv. Since the das_client is called from within a script which
parses sys.argv for command line arguments, calling a subprocess with a
set list of arguments avoids setting and restoring sys.argv manually.