Modify das_query options, start writing prep.py
The --outdir option of das_query has been removed. Not only will this
option be used all the time, but a consistent place to specify it is
needed to propagate it to other scripts as needed. It has since been
promoted to a configuration file option in "samples_nosplit.ini"

A first pass at a rework of treecopier has been started, with an
accompanying prep.py script for running the preparation step in the
context of lxplus.